ingestion/requirements.txt
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
pyyaml>=6.0.0
